+++ src/main/kotlin/com/atlassian/performance/tools/jvmtasks/api/Backoff.kt
@@ -4,12 +4,15 @@ import java.time.Duration
 
 /**
  * Calculates the waiting time for a given action attempt.
+ *
+ * @since 1.0.0
  */
 interface Backoff {
 
     /**
      * @param attempt Counts action attempts.
      * @return Separates action attempts in time.
+     * @since 1.0.0
      */
     fun backOff(
         attempt: Int